validation/requirements.txt
playwright
playwright-stealth
motor
pydantic
python-dotenv